Implemented common component registry in client
